Move Asset filter component TO MediaGalleryUi module
