+++ MediaGalleryUi/view/adminhtml/web/js/form/element/ui-select.js
