+++ MediaGalleryUi/view/adminhtml/web/template/grid/filters/elements/ui-select.html
